Remove effects test file and remove unused refx dependency
